backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0
openai>=1.40.0
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
aiofiles==23.2.1
httpx>=0.27.0
